Extract vehicle-to-route projection from route-detail.store
The store carried ~180 lines of pure geometry — projecting a live vehicle onto
a route shape, bracketing it between stops and deriving travel direction from
bearing — with no test coverage. It moves to lib/transit/route-projection.ts
with 14 tests, and the route-detail domain types move to types/transit.types
so the lib module does not import upward from a store.

route-detail.store.ts: 838 -> 646 lines.
