+++ web/src/services/layers/features/route-isolation.service.ts
@@ -11,7 +11,8 @@
  */
 
 import { watch, type WatchStopHandle } from 'vue'
-import { useRouteDetailStore, type RouteDetailStop } from '@/stores/route-detail.store'
+import type { RouteDetailStop } from '@/types/transit.types'
+import { useRouteDetailStore } from '@/stores/route-detail.store'
 import { densifyLine } from '@/lib/geo/geo-densify'
 import { projectAlong, sliceAlong } from '@/lib/geo/geo-line'
 import { widthExpr } from '@/services/layers/features/portolan/portolan-expressions'
