+++ web/src/stores/route-detail.store.test.ts
@@ -6,8 +6,9 @@
  */
 
 import { describe, test, expect, beforeEach, vi } from 'vitest'
+import type { RouteDetailStop, VehicleOnRoute } from '@/types/transit.types'
 import { setActivePinia, createPinia } from 'pinia'
-import { useRouteDetailStore, type RouteDetailStop, type VehicleOnRoute } from '@/stores/route-detail.store'
+import { useRouteDetailStore } from '@/stores/route-detail.store'
 import type { TransitVehiclePosition } from '@/types/multimodal.types'
 
 // Mock the API and realtime modules
